Move ChangeView to TS
